requirements.txt
python-telegram-bot[webhooks]==20.7
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2
joblib==1.3.2
transformers==4.35.2
torch==2.1.1
newsapi-python==0.2.7
SQLAlchemy==2.0.23
matplotlib==3.8.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
